requirements.txt
torch
torchvision
numpy
Pillow
tqdm
scikit-learn
matplotlib
